Moved get_signed_claims_payload() to mock
