+++ pallets/handles/src/tests/handle_retirements_tests.rs
@@ -25,6 +25,5 @@ fn retire_handle_no_handle() {
 			),
 			Error::<Test>::MSAHandleDoesNotExist
 		);
-
 	});
 }
